Bake in pprof data from production to optimise hot paths
This data was collected by setting `COLLECT_METRICS=yes` on production and then running:

```bash
$ while true; do curl http://localhost:6060/debug/pprof/profile?seconds=30 --output profiles/$(date +%s).pprof; sleep 270; done
```

To gather a chunk of metrics every 5 mins for 30 seconds for a couple of hours. The traffic to the server came in bursts, but included lots of visits to jupyter- which is massively asset heavy and makes loads of calls to /api/v1/auth.

At some point we can start thinking about gathering data from loads of instances over many days with lots of requests; at that point this commit will be useful prep