Remove dependency on pprof files
Because we don't necessarily push these files up to github we see
failures in CI; to wit:

```
make: *** No rule to make target 'profiles/*.pprof', needed by 'merged.pprof'.  Stop.
```

We should keep the dependency on `merged.pgo`  in making the binary, since that
requires the merged profiles and because we want to update the binary
when that changes, but remove the other